Refactored workflowDataSource class
