+++ src/app/workflow-maint/workflow-maint.component.ts
@@ -7,6 +7,7 @@ import { Observable } from 'rxjs/Observable';
 
 import { WorkflowEntry } from 'app/view-models/workflowEntry';
 import { WorkflowEntryService } from 'app/services/workflow-entry.service';
+import { WorkflowDataSource } from 'app/workflow-maint/workflowDataSource';
 
 @Component({
   selector: 'toil-workflow-maint',
@@ -30,47 +31,6 @@ export class WorkflowMaintComponent implements OnInit {
   createEntry() {
     this.router.navigate(['/authenticated/workflow-detail', 0, 'create']);
   }
-
 }
 
-export class WorkflowDataSource extends DataSource<any> {
 
-  constructor(private workflowEntrySVC: WorkflowEntryService, private sort: MdSort) {
-    super();
-  }
-
-  connect(): Observable<WorkflowEntry[]> {
-    const displayDataChanges = [
-      this.workflowEntrySVC.dataChange,
-      this.sort.mdSortChange
-    ];
-
-    return Observable.merge(...displayDataChanges).map(() => {
-      return this.getSortedData();
-    });
-  }
-
-  disconnect() { }
-
-  getSortedData(): WorkflowEntry[] {
-    const data = this.workflowEntrySVC.dataChange.value;
-
-    if (!this.sort.active || this.sort.direction === '') {
-      return data;
-    }
-
-    return data.sort((a, b) => {
-      let propertyA: number | string = '';
-      let propertyB: number | string = '';
-
-      switch (this.sort.active) {
-        case 'name': [propertyA, propertyB] = [a.name, b.name]; break;
-      }
-
-      const valueA = isNaN(+propertyA) ? propertyA : +propertyA;
-      const valueB = isNaN(+propertyB) ? propertyB : +propertyB;
-
-      return (valueA < valueB ? -1 : 1) * (this.sort.direction === 'asc' ? 1 : -1);
-    });
-  }
-}
